Changes the fetching of tiles
Previously, the image collection would create a mosaic and then clip the required area. The issue revolved around the mosaic having holes and those holes being clipped into the tile. Another issue is there is no properties for the tile when the mosaic is created. So by filtering for tiles that contain the area we retain all of the properties surrounding that tile as well as detecting areas that do not have data for the given constraints. Also added some extra folders for the data directory.